refactor note into delete and save packages
