+++ src/use_case/note/delete/DeleteNoteOutputBoundary.java
@@ -1,4 +1,4 @@
-package use_case.note;
+package use_case.note.delete;
 
 public interface DeleteNoteOutputBoundary {
