+++ src/use_case/note/save/SaveNoteInputBoundary.java
@@ -1,4 +1,4 @@
-package use_case.note;
+package use_case.note.save;
 public interface SaveNoteInputBoundary {
     void execute(SaveNoteInputData saveNoteInputData);
